app/requirement.txt
alembic==1.5.6
aniso8601==9.0.1
appdirs==1.4.4
astroid==2.5.1
attrs==20.3.0
bcrypt==3.2.0
blinker==1.4
certifi==2020.12.5
cffi==1.14.5
colorama==0.4.4
config==0.5.0.post0
crontab==0.22.9
distlib==0.3.1
filelock==3.0.12
flasgger==0.9.5
Flask
Flask-Bcrypt==0.7.1
Flask-JSON==0.3.4
Flask-Mail==0.9.1
Flask-Migrate==2.7.0
Flask-RESTful==0.3.8
flask-restful-swagger==0.20.2
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
flask-swagger-ui==3.36.0
Flask-Testing==0.8.1
isort==5.7.0
Jinja2
jsonschema==3.2.0
lazy-object-proxy==1.5.2
logger==1.4
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
pbr==5.5.1
pdoc3==0.9.2
pipenv==2020.11.15
priority==1.3.0
psycopg2-binary
pycparser==2.20
PyJWT==2.0.1
pylint==2.7.2
pyrsistent==0.17.3
python-crontab==2.5.1
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
PyYAML==5.4.1
six==1.15.0
SQLAlchemy==1.3.23
stevedore==3.3.0
toml==0.10.2
virtualenv==20.4.2
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
virtualenvwrapper.basic==0.2.2
Werkzeug==0.16.1
wincertstore==0.2
wrapt==1.12.1
